Display dstchannel column
It's a rather common case in Asterisk to have a Dial calling
multiple destinations:

exten => _1234567,n,Dial(SIP/101&SIP/102&SIP/103)

Asterisk 12+ creates multiple records for the call, all of them
having "1234567" in dst, so having only src and dst it's impossible
to figure out which of the SIP phones answered.

The only column that has this information is dstchannel,
so it is reasonable to have in the table by default.

Closes #6